# Avoid allocations for Unicode data tries