… src/Avalonia.Base/Media/TextFormatting/Unicode/BiDi.trie.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/GraphemeBreak.trie.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/UnicodeData.trie.cs
… tests/Avalonia.Base.UnitTests/Media/TextFormatting/UnicodeDataGenerator.cs