## What does the pull request do?
This PR completely avoids allocations for Unicode data tries on .NET 7+, removing 195 kB of always allocated managed memory.

## How was the solution implemented (if it's not obvious)?
The trie's data is now a property returning `ReadOnlySpan<uint>` instead of `ReadOnlySpan<byte>`, that's used directly instead of copying the bytes to a runtime `uint[]`.

Starting with .NET 7 and Roslyn 17.5, this syntax makes the compiler use [`RuntimeHelpers.CreateSpan`](https://learn.microsoft.com/en-us/dotnet/api/system.runtime.compilerservices.runtimehelpers.createspan?view=net-8.0), which references the data embedded inside the assembly directly on little-endian systems, without having to allocate or copy anything!
(For .NET Standard, the compiler falls back to an array.)

## Performance
I expected better memory usage, but it turns out we also get some performance improvement. With these changes, the JIT is able to inline some constants from the Unicode data directly, generating better code.

Before:
| Method               | CodepointValue | Mean      | Error     | StdDev    | Allocated |
|--------------------- |--------------- |----------:|----------:|----------:|----------:|
| GetGeneralCategory   | 65             | 0.4334 ns | 0.0118 ns | 0.0110 ns |         - |
| GetGeneralCategory   | 4660           | 0.4076 ns | 0.0215 ns | 0.0201 ns |         - |

After:
| Method               | CodepointValue | Mean      | Error     | StdDev    | Allocated |
|--------------------- |--------------- |----------:|----------:|----------:|----------:|
| GetGeneralCategory   | 65             | 0.2274 ns | 0.0101 ns | 0.0089 ns |         - |
| GetGeneralCategory   | 4660           | 0.2292 ns | 0.0056 ns | 0.0050 ns |         - |