@maxkatz6 2024-03-21
> Starting with .NET 7 and Roslyn 17.5, this syntax makes the compiler use [RuntimeHelpers.CreateSpan](https://learn.microsoft.com/en-us/dotnet/api/system.runtime.compilerservices.runtimehelpers.createspan?view=net-8.0), which references the data embedded inside the assembly directly on little-endian systems, without having to allocate or copy anything!

I am pretty sure we expected old code to work exactly in this way, i.e. returning reference to the data inside of the assembly. I.e. for .NET 6 as well.
But I guess it was wrong.