@MrJul 2024-03-22
> I am pretty sure we expected old code to work exactly in this way, i.e. returning reference to the data inside of the assembly. I.e. for .NET 6 as well.
> But I guess it was wrong.

You were right, it's been working for a while with `ReadOnlySpan<byte>`, but the old code then copied that to a runtime `uint[]`, so we still had allocations.

Supporting other primitive types than `byte` with `RuntimeHelpers.CreateSpan` is much more recent: https://github.com/dotnet/roslyn/pull/61414